refactor: refactor on AccountView
